utils/requirements.txt
# Python dependencies for MBD-Website utils
# To install: pip install -r requirements.txt

# Google Drive API
google-api-python-client==2.126.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# For API requests
requests==2.31.0

# Other common utilities
python-dotenv==1.0.0 
